requirements.txt
requests
django==2.2.16
djangorestframework==3.12.4
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
django-filter==2.4.0
djangorestframework-simplejwt==4.7.2
python-dotenv==0.19.0